backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
beautifulsoup4==4.12.2
openai==1.3.7
moviepy==1.0.3
Pillow==10.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
huggingface_hub[inference]
